+++ guillotina/db/transaction.py
@@ -4,6 +4,7 @@ from guillotina.db.interfaces import IStorageCache
 from guillotina.db.interfaces import ITransaction
 from guillotina.db.interfaces import ITransactionStrategy
 from guillotina.db.interfaces import IWriter
+from guillotina.db.oid import generate_oid
 from guillotina.db.reader import reader
 from guillotina.exceptions import ConflictError
 from guillotina.exceptions import ReadOnlyError
@@ -20,7 +21,6 @@ import asyncio
 import logging
 import sys
 import time
-import uuid
 
 
 _EMPTY = '__<EMPTY VALUE>__'
@@ -235,13 +235,12 @@ class Transaction(object):
             if new_oid is not None:
                 new = True
             else:
-                new_oid = uuid.uuid4().hex
+                new_oid = generate_oid(obj)
             oid = new_oid
 
         obj._p_oid = oid
         if new or obj.__new_marker__:
             self.added[oid] = obj
-            obj.__new_marker__ = False
         elif oid not in self.modified and oid not in self.added:
             self.modified[oid] = obj
 
@@ -372,6 +371,7 @@ class Transaction(object):
         await self._strategy.tpc_commit()
         for oid, obj in self.added.items():
             await self._store_object(obj, oid, True)
+            obj.__new_marker__ = False
         for oid, obj in self.modified.items():
             await self._store_object(obj, oid)
         for oid, obj in self.deleted.items():