+++ CHANGELOG.rst
@@ -1,7 +1,11 @@
 3.0.9 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Use unique constraint instead of doing lookups on existing content for inserts
+  [vangheem]
+
+- On content creation, use summary serialization instead of full object serialization
+  [vangheem]
 
 
 3.0.8 (2018-04-06)